requirements.txt
aiofiles==0.6.0
asgiref==3.3.1
b0mb3r==3.1.3
beautifulsoup4==4.9.1
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cx-Freeze==6.3
cx-logging==2.2
cycler==0.10.0
dataclasses-json==0.5.2
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
docopt==0.6.2
et-xmlfile==1.0.1
fastapi==0.61.2
google==3.0.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
gunicorn==20.1.0
h11==0.11.0
httpcore==0.12.1
httplib2==0.18.1
httpx==0.16.1
idna==2.10
importlib-metadata==2.0.0
instaloader==4.6.2
isodate==0.6.0
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.3.1
loguru==0.5.3
lxml==4.6.2
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
matplotlib==3.3.3
monotonic==1.5
mypy-extensions==0.4.3
networkx==2.5
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.4
pandas==1.1.0
phonenumbers==8.12.13
Pillow==8.0.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.2
PyDrive==1.3.1
pygdrive3==0.7.0
pyparsing==2.4.7
pyperclip==1.8.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
PySocks==1.7.1
pyTelegramBotAPI==3.7.6
python-a2s==1.3.0
python-dateutil==2.8.1
python-valve==0.2.1
pytz==2020.4
PyYAML==5.3.1
random-user-agent==1.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.7
selenium==3.141.0
sentry-sdk==0.19.3
six==1.15.0
sniffio==1.2.0
socks==0
soupsieve==2.2
spark-parser==1.8.9
sqlparse==0.4.1
starlette==0.13.6
stringcase==1.2.0
telebot==0.0.4
termcolor==1.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uncompyle6==3.7.4
uritemplate==3.0.1
urllib3==1.25.11
uvicorn==0.12.2
valve==0.0.0
webdriver-manager==3.2.2
whitenoise==5.2.0
win32-setctime==1.0.3
xdis==5.0.7
xlrd==2.0.1
zipp==3.4.0
